Update currency.lua
I fixed my issue where my function was not being called correctly. It will now be called on xpgains and shares the event watcher as the "refresh" function. 

Limitations to note: The calculator requires 2 kills to estimate the number of kills remaining and it is only a simple subtraction between the last 2 xp gains divided by the remaining xp necessary.